requirements.txt
charset-normalizer==3.3.2
numpy==1.26.4
python-dateutil==2.9.0.post0
requests==2.32.2
schedule==1.2.2
urllib3==2.2.1